#16 Rethink testing strategy
There is a new boot task called `slacker-test`, which uses the excellent
boot-test library, and which runs all `deftest` functions found. The switch
to regular clojure.test is made tolerable primarily because boot-test makes a
good effort to present human-readable diffs in expected and actual input -
not entirely unlike expectations before it.

This rewrites a lot of lines, but shouldn't touch on anything dangerous.