Emit event when connect-bot fails
#49

When connect-bot fails due to IO, incorrect credentials or otherwise,
the error only appears in the log, which makes it unnecessarily hard to
response to. This sort of error should really be handled through a
well-defined, general mechanism for errors, but as this doesn't exist
yet, I'm resorting to just emitting this simple event for now. That, at
least, allows users to respond to the failure.

It seems likely that I'll change my mind on how, exactly, errors should
be handled in the general case.